backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
pydantic==2.6.1
sqlalchemy==2.0.39
jieba==0.42.1
pypinyin==0.50.0
python-multipart==0.0.9
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
loguru==0.7.2